Move to React 0.14 and some other fixes
