+++ example/package.json
@@ -15,7 +15,8 @@
     "express-graphql": "^0.3.0",
     "graphql": "^0.4.4",
     "multer": "^1.0.3",
-    "react": "^0.13.3",
+    "react": "^0.14.2",
+    "react-dom": "^0.14.2",
     "react-redux": "^0.8.2",
     "react-router": "1.0.0-beta3",
     "redux": "^1.0.0-rc",
