+++ example/src/client/graphql/index.js
@@ -2,8 +2,10 @@
 
 import 'whatwg-fetch';
 import React from 'react';
+import ReactDOM from 'react-dom';
 import { Router } from 'react-router';
 import { history } from 'react-router/lib/BrowserHistory';
+
 import routes from './routes';
 import { Adrenaline, createGraphQLAdaptor } from '../../../../src';
 import schema from 'shared/schema';
@@ -11,9 +13,9 @@ import schema from 'shared/schema';
 const adaptor = createGraphQLAdaptor('/graphql', schema);
 
 const rootNode = document.getElementById('root');
-React.render(
+ReactDOM.render(
   <Adrenaline adaptor={adaptor}>
-    {() => <Router history={history} children={routes} />}
+    <Router history={history} children={routes} />
   </Adrenaline>,
   rootNode
 );
