--- src/components/Loading.jsx
@@ -1,11 +1,0 @@
-/* @flow */
-
-import React, { Component } from 'react';
-
-export default class Loading extends Component {
-  render() {
-    return (
-      <div>Loading data...</div>
-    );
-  }
-}
